requirements.txt
fastapi
uvicorn[standard]
google-generativeai>=0.8.5
pydantic
